+++ core/oram_controller.h
@@ -14,8 +14,8 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-#ifndef ORAM_IMPL_CLIENT_ORAM_CONTROLLER_H_
-#define ORAM_IMPL_CLIENT_ORAM_CONTROLLER_H_
+#ifndef ORAM_IMPL_CORE_ORAM_CONTROLLER_H_
+#define ORAM_IMPL_CORE_ORAM_CONTROLLER_H_
 
 #include <grpc++/grpc++.h>
 
@@ -88,176 +88,6 @@ class OramController {
   }
 };
 
-// A trivial solution for ORAM. It only has educational meaning. Do not use it
-// in any productive environment.
-class LinearOramController : public OramController {
- private:
-  grpc::Status ReadFromServer(std::string* const out);
-  grpc::Status WriteToServer(const std::string& input);
-
- protected:
-  virtual OramStatus InternalAccess(Operation op_type, uint32_t address,
-                                    oram_block_t* const data,
-                                    bool dummy = false) override;
-
- public:
-  LinearOramController(uint32_t id, bool standalone, size_t block_num)
-      : OramController(id, standalone, block_num, OramType::kLinearOram) {}
-  virtual OramStatus InitOram(void) override;
-  virtual OramStatus FillWithData(
-      const std::vector<oram_block_t>& data) override;
-};
-
-class SquareRootOramController : public OramController {
-  // The layout of the square root ORAM is:
-  // ------------------ | ------- | -------|
-  //    m main sotrage    sqrt(m)   sqrt(m)
-  //                      dummy     shelter
-  // |       permuted memory      |
-  //
-  // Select a permutation \pi over the integers 1, ..., m + \sqrt{m} and
-  // obliviously relocate the words according to the permutation.
-  sqrt_oram_storage_t shelter_;
-
- protected:
-  virtual OramStatus InternalAccess(Operation op_type, uint32_t address,
-                                    oram_block_t* const data,
-                                    bool dummy = false) override;
-
- public:
-  SquareRootOramController(uint32_t id, bool standalone, size_t block_num);
-};
-
-// This class is the implementation of the ORAM controller for Path ORAM.
-class PathOramController : public OramController {
-  friend class PartitionOramController;
-
-  uint32_t id_;
-  // ORAM parameters.
-  uint32_t tree_level_;
-  uint8_t bucket_size_;
-  uint32_t number_of_leafs_;
-  // stash size.
-  size_t stash_size_;
-
-  p_oram_position_t position_map_;
-  // The stash should be tied to the slots of Partition ORAM, so we use
-  // pointers to manipulate the stash.
-  p_oram_stash_t stash_;
-  // Networking time.
-  std::chrono::microseconds network_time_;
-  // Networking communication.
-  size_t network_communication_;
-
-  // ==================== Begin private methods ==================== //
-  OramStatus ReadBucket(uint32_t path, uint32_t level,
-                        p_oram_bucket_t* const bucket);
-  OramStatus WriteBucket(uint32_t path, uint32_t level,
-                         const p_oram_bucket_t& bucket);
-  OramStatus AccurateWriteBucket(uint32_t level, uint32_t offset,
-                                 const p_oram_bucket_t& bucket);
-  OramStatus PrintOramTree(void);
-
-  p_oram_stash_t FindSubsetOf(uint32_t current_path);
-  // ==================== End private methods ==================== //
- protected:
-  virtual OramStatus InternalAccess(Operation op_type, uint32_t address,
-                                    oram_block_t* const data,
-                                    bool dummy = false);
-  // The separate interface for reading is reserved for direct designation of
-  // position, which is useful to ODS.
-  virtual OramStatus InternalAccessDirect(Operation op_type, uint32_t address,
-                                          uint32_t position,
-                                          oram_block_t* const data,
-                                          bool dummy = false);
-
- public:
-  PathOramController(uint32_t id, uint32_t block_num, uint32_t bucket_size,
-                     bool standalone = true);
-
-  virtual OramStatus InitOram(void) override;
-  virtual OramStatus FillWithData(
-      const std::vector<oram_block_t>& data) override;
-  virtual uint32_t RandomPosition(void) override;
-
-  virtual OramStatus AccessDirect(Operation op_type, uint32_t address,
-                                  uint32_t position, oram_block_t* const data) {
-    return InternalAccessDirect(op_type, address, position, data, false);
-  }
-
-  uint32_t GetTreeLevel(void) const { return tree_level_; }
-  size_t ReportClientStorage(void) const;
-  size_t ReportStashSize(void) const { return stash_size_; }
-  size_t ReportNetworkCommunication(void) const;
-  std::chrono::microseconds ReportNetworkingTime(void) const {
-    return network_time_;
-  }
-};
-
-// This class is the implementation of the ORAM controller for Partition ORAM.
-class PartitionOramController final : public OramController {
-  size_t partition_size_;
-  size_t bucket_size_;
-  size_t nu_;
-  static size_t counter_;
-  // Position map: [key] -> [slot_id].
-  p_oram_position_t position_map_;
-  // Slots: [slot_id] -> [block1, block2, ..., block_n].
-  pp_oram_slot_t slots_;
-  // Controllers for each slot: [slot_id] -> [controller_1, controller_2, ...,
-  //                                          controller_n].
-  // TODO: Currently this is pathoram controller.
-  std::vector<std::unique_ptr<PathOramController>> path_oram_controllers_;
-
-  PartitionOramController(uint32_t id = 0ul)
-      : OramController(id, true, 0ul, OramType::kPartitionOram) {}
-
-  // ==================== Begin private methods ==================== //
-  OramStatus Evict(uint32_t id);
-  OramStatus SequentialEvict(void);
-  OramStatus RandomEvict(void);
-
-  OramStatus ProcessSlot(const std::vector<oram_block_t>& data,
-                         uint32_t slot_id);
-  // ==================== End private methods ==================== //
- protected:
-  virtual OramStatus InternalAccess(Operation op_type, uint32_t address,
-                                    oram_block_t* const data,
-                                    bool dummy = false) {
-    UNIMPLEMENTED_FUNC;
-  }
-
- public:
-  static std::unique_ptr<PartitionOramController> GetInstance();
-
-  void SetBucketSize(size_t bucket_size) { bucket_size_ = bucket_size; }
-  void SetNu(size_t nu) { nu_ = nu; }
-
-  virtual OramStatus Access(Operation op_type, uint32_t address,
-                            oram_block_t* const data) override;
-
-  virtual OramStatus FillWithData(
-      const std::vector<oram_block_t>& data) override;
-  virtual OramStatus InitOram(void) override;
-
-  OramStatus Run(uint32_t block_num, uint32_t bucket_size);
-  // A reserved interface for testing one of the PathORAM controllers.
-  OramStatus TestPathOram(uint32_t controller_id);
-  OramStatus TestPartitionOram(void);
-
-  size_t ReportClientStorage(void) const;
-  size_t ReportNetworkCommunication(void) const;
-  std::chrono::microseconds ReportNetworkingTime(void) const;
-
-  void Reset(uint32_t block_num) {
-    block_num_ = block_num;
-    position_map_.clear();
-    slots_.clear();
-    path_oram_controllers_.clear();
-  }
-
-  virtual ~PartitionOramController() {}
-};
 }  // namespace oram_impl
 
-#endif  // ORAM_IMPL_CLIENT_ORAM_CONTROLLER_H_
+#endif  // ORAM_IMPL_CORE_ORAM_CONTROLLER_H_
